UI fixturre

+++ Source/FMTspatialschedule.cpp
@@ -841,7 +841,9 @@ std::vector<FMTcoordinate> FMTSpatialSchedule::GetGroupsConflict(const Core::FMT
 				case Core::FMTconstrainttype::FMTspatialsize:
 						{
 						returnvalue += m_events.EvaluateSize(ACTION_USED,
-							period, lower, upper, TEST_LOWER_BOUND);
+							period, 
+							static_cast<size_t>(lower), static_cast<size_t>(upper),
+							TEST_LOWER_BOUND);
 						break;
 						}
 					case Core::FMTconstrainttype::FMTspatialadjacency:
